CSN: CSN-hasim-193
Updated the traffic light example to work with the new fpgaenv structure.
To work reasonably in simulation the traffic light delay was made
a paramater and set very short for the software simulation versions.

Updated hasim-connect to include module parameters for the build
of the module at the synthesis boundary.