CSN: CSN-hasim-207
1) Local Controller stub is in place. Currently it starts the program via scanchain and returns when done. It
_could_ control the Ports except that the Ports just always say that they are ready and balanced.

2) Makefile now returns 1 correctly if a subdirectory fails. It turns out that the for-loop in the Makefile
was a bash for-loop rather than a Makefile for-loop. This confusion meant that the for-loop returned
"success" [0] even if a child process had failed [1]. We now test for this explicitly and bottom-out
correctly. This might fix some of the weirdness people had experienced at synthesis boundaries with
connections, as these rely on logfiles which might have gotten confused by the incorrect bottoming-out.

Michael P.