CSN: CSN-hasim-299
Added a first pass at assertions. The assertions interface is in place, but uses the scanchain.

Assertions are very "imprecise" in that it may be many FPGA cycles before we
notice an assertion has been raised. Additionally, if more than one assertion is
raised then it is possible that we will report a derivative assertion rather
than the one that was the root cause. There's probably no way to avoid this
without changing the interconnect topology (a longterm project.)

Michael P.