CSN: CSN-hasim-324
Put generated Bluespec interface include files in asim/provides tree.

For now you can include these interface files either with the new style:

`include "asim/provides/foo.bsh"

or the old style:

`include "foo.bsh"

This works because both the root of the include tree and the leaf are
on the include search path.  When we are ready we can drop the leaf
from the search path by editing SCons.hw.template.