CSN: CSN-hasim-330
Dictionaries are now hierarchical.  Dictionary A may have a sub-dictionary
B using the syntax:

   def A.B.FOO "foo string";

Breaking up the dictionary allows modules to define their own sub-dictionaries
of a larger category.  The sub-dictionaries share the same numbering space.
Including the header file at a given level in the tree automatically imports
all definitions below the level.  With this mechanism recompilation due
to dictionary changes can be minimized.

In C++ the name of the printing class changed from DICT_<NAME> to <NAME>_DICT.

In Bluespec the elements are now preprocessor definitions, e.g.
     `define STREAMID_ASSERT 0
The Bluespec ID to string function is now in a module named <NAME>_DICT and
the function is named show<NAME>_DICT.