CSN: CSN-hasim-346
Updated the hasim platform interface to use the front panel
and streams virtual devices with individual arguments
to the remote methods rather than passing a struct as
the only argument.

Also updated the pure Bluespec demos to do the same. Added
a demo that uses both the streams, and front panel devices.
And did a re-arrangement of the bluespec demo models.