CSN: CSN-hasim-363
Version 3 of the Functional Partition now supports an isa-emulator service via
RRR.

Upon encountering an emulated instruction, the current value of all registers is
passed to the software via the RRR syncRegister() call. Then a call to
emulateInstruction() is made. The software makes one or more calls to
updateRegister(), and finally a call to emulationFinished(). (This last call
will eventually be replaced with an ACK.)

Removed some exraneous instantiations from the controller.

Michael P.