CSN: CSN-hasim-368
I believe there was a type in regstate_manager_macro_ops.bsv, line 987.
I also changed the ISA_REG_INDEX to a bit-vector type because it uses maxBound, and is used in array indexing, both of which can't be done easily with tagged unions. I will be changing the return type of getIsaSrc and getIsaDst to return Bits instead.