CSN: CSN-hasim-398
Tired of redundant sizes sprinkled through the functional partition
configurations.  Created one, top level, "funcp_base_types" requirement.
Files in the base types must be terminal within the functional partition.
For now defined here is the ISA address size and the integer register
size.  From these other parts of the functional partition pick sizes
for funcp_memory and isa_emulator.  The only remaining other place where
a size is needed is the memory_store_info_size in funcp_memory, due to
temporary RRR limitations.  The sw side of the memory interface verifies
that memory_store_info_size is set correctly.

Also changed the protocol for ISA emulator to return a branch.  Using
the high bit caused problems since we can't know where the high bit is
without assuming a PC size.  Since the branch bit as part of the returned
PC is a temporary hack I used bits 1:0, which must be 0 for both Alpha and
MIPS.  Of the 4 states, 0 is no branch, 1 is branch, 2 is unused and 3
means terminate.  The exit status is in bit 2 for terminate.