CSN: CSN-hasim-415
Created a separate module for command line switch processing. That
module is used by main and %required by the hybrid controllers. A
controller was added to the pure bluespec models so they can share the
switch parsing. Those models are currently using the full hasim switch
parser, which is overkill but works. 

Moved the %library specification of libasim to hasim_common to match
the convention in the the pure bluespec designs.

Added a  regression run for the base hasim package. This benchmark
currently builds but does not run properly.