CSN: CSN-hasim-416
Switched the null benchmark to use the standard setup-bmark
script and modified that script to handle the 'none' feeder
and the run.bmark script to create a fake stats file for that
feeder as well.

Now the 'hello world' hybrid demo runs and passes as a 
regression. So in your asimrc file one can test the hasim,
hasim-alpha, and hasim-mips package with the lines:

[regression.launcher]
PACKAGES=hasim,hasim-alpha,hasim-mips