CSN: CSN-hasim-435
Get rid of the warnings about conflicting writes to the prfValids vector
by adding an execution_order pragma.  According to Bluespec this does not
affect the scheduler.  It only affects the priority encoder for outputs
of a cycle.  When the warning is printed the compiler picks a priority
anyway, so there is no problem using the pragma.  There is a longer
discussion at http://bluespec.com/forum/viewtopic.php?p=281