CSN: CSN-hasim-437
Increased size of functional partition cacheline. Previously it was set
to be equal to the word-size. The `FUNCP_CACHLINE_BITS parameter
indicates the size of the unctional partition cacheline in bits. Since
this value must be known external to the cache as well, the parameter is
defined in hybrid-funcp-memory.awb.

With cachelines bigger than words, a store sent to the cache that misses
must be translated into a load to memory.

The valid bits of the cache were not explicitly initialized to zero; fixed.

Passes hello and memory_tests_aligned in bluesim.