CSN: CSN-hasim-497
Modified the hasim build scripts to allow multiple AWB files to declare BDPI imported functions. The original technique was to use %param  BDPI_NAMES as a magic parameter that would cause the correct list of .ba files to be known to the SConscript. Now, we can use any identifier that _begins_ with the string BDPI_NAMES, and the values get appended.