CSN: CSN-hasim-501
Top level mkModel now instantiates a clock generator to feed the full hasim
system. %param MODEL_CLOCK_FREQ is used to set the frequency at which the
model would run. Only certain frequencies are legal, and if you use one that
is not, then the bluespec compilation will bail out. (This functionality is
brought to you by fpga-components.)