CSN: CSN-hasim-546
Split the register state manager, which had grown to 3500 lines, into separate
files.  Individual functional/timing interface pipelines are in files named
regmgr-macro-pipe-<interface>.  This first commit generates equivalent Verilog
to the previous version.  All global state that used to be at the top of the
large file is now declared in regmgr-macro.bsv and passed in to the constructor
for each pipeline.

The next step will be to reduce the global state and relax inter-pipeline
scheduling constraints.  Simplifying the schedule will, hopefully,
allow for higher FPGA frequencies.