CSN: CSN-hasim-579
The Bluespec compiler's scheduler appears to have exponential compile-time
performance degradation when doing a lookup on a vector of 4-way tiny caches.
Clearly this can be generalized, and Ravi at Bluespec is working on that.
In the mean time the computed scheduler predicates can be radically simplified
for some rules using the conservative_implicit_conditions pragma.  This pragma
simplifies the predicate by, among other things, requiring that all input and
output FIFOs be ready even if they are accessed conditionally in the rule.

The specific changes:
  - Use conservative_implicit_conditions in both the HAsim cache code and in
    the TLB cache.  Some rules in the TLB had to be split in order to avoid
    deadlocks.
  - Change build script to generate Bluespec scheduler dump (.sched files)
    when building wrappers.

Simplification of predicates for just these 6 rules reduced the LUT usage of
the unpiplined model by 3% with no changes in timing and reduced compile time
of the TLB synthesis boundary by about 90%.