CSN: CSN-hasim-586
Another step toward allowing the HASIM_CACHE implementation to return results
OOO.  Read responses used to be just the data.  Now read responses include
the original address and caller-defined metadata associated with the read
request.  With this there will be no need for a ROB to associate details
about read requests with the read data coming from the cache.