CSN: CSN-hasim-661
Add a new soft connection for passing register writebacks from the ISA datapath
to the register state manager.  This decouples register writes from the response
from getResults to the timing partition.  The functional model blocks on attempts
to read registers that are not yet written.

The end of the getResults pipeline is now much simpler since it doesn't have
to cope with multiple register write requests all arriving in the same cycle.
The bus width of the soft connection is now a single register instead of the
maximum number of instruction destinations.