CSN: CSN-hasim-782
Add a merge buffer for writes and merge sequential read misses for accesses
to the same uncached scratchpad word.  This reduces traffic on the scratchpad
ring.  The scratchpad virtual device in the platform already has similar buffers
for each uncached scratchpad at the line level to reduce I/O to the host.