CSN: CSN-hasim-800
Add a 16-wide version of the channel integrity test with 8 random values followed
by 8 complements.  This gives more freedom for patterns on ACP, since the ACP
data-path is 256 bits wide.

The 16-wide test is a separate set of sources because of the lack of
support for vector types in RRR.