CSN: CSN-hasim-807
SCons 1.2 isn't building synthesis boundary wrappers after running hasim-connect
if the wrapper is unchanged.  That is probably correct, but Bluespec complains
due to dependence timestamp problems.  For now, force the generated wrapper to
be different whenever it is built by adding the current time.  This only
happens when the wrapper is rebuilt due to other source changes, so it
shouldn't force extra builds.