CSN: CSN-hasim-818
The following large checkin integrates soft services into hasim
connect while maintaining support for existing hasim connect based
applications.  To use the sofft services, an application must be
switched to the soft services type, use the soft service apm, and be a
pipeline build application.  Otherwise, things won't compile.  Most of
the reason for this non-compilation is to protect existing builds from
the new code.  

To prevent bit rot, the traffic light codes now use the
build pipeline and soft services.  I have tested this to ensure that a
functional bit-file is produced on the ACP.