CSN: CSN-hasim-831
Two sets of modifications - some inexorable groundwork for multifpga
builds (in the sense that I couldn't unedit the files).  However,
these changes are ignored by downstream tools, and do not affect code
paths using the non-pipelined build.  There are some minor changes to
the hasim scripts.  Most of these are generalizations of code which
broke once I started building things from the root of the source tree,
as opposed to intervening nodes.  I also add an option for inserting a
hash into the generated verilog stub files.  This prevents unecessary
rebuilds when the corresponding wrappers change, but do not modify
thier interface.