CSN: CSN-hasim-855
Fix our horrible implementation of buffered A-ports.  The buffer is essentially
a FIFO, but was hand implemented with a depth equal to that needed for maximum
latency of all ports, in spite of knowing the true latency of a given port.

Changes:

- Switch to forwarding data through normal FIFOs.
- Heuristically choose between FIFO in distributed memory (mkSizedFIFO) and
  a BRAM storage (mkSizedBRAMFIFO).  Our BRAM FIFO has a bypass so the timing
  of the two is identical.

Now the only object sized by PORT_MAX_LATENCY is the index used during
initialization.

Before the change, 30% of distributed memory was allocated unnecessarily in
the 64-way in-order mesh model, making it impossible to meet reasonable timing.