CSN: CSN-hasim-862
The functional register file was a complicated algorithm with 2 read ports
to support reading both operands of an instruction in parallel.  The 2nd read
port doubled the required BRAM, which becomes a problem as we add
contexts.  Switch to a 4 banked memory, with one read port per bank.
In addition to taking half the BRAM, the code is much simpler.