CSN: CSN-hasim-887
Add experimental soft connection compression instances for A-Ports that cross
multi-FPGA boundaries.  The code currently special cases names of connections
from the in-order mesh model when inorder_pipeline_front_end and decode_stage
are moved to ACP0 (normally on ACP1).

Two compressors are implemented:  one simply eliminates the t_MSG from
Maybe#(t_MSG) when the predicate is false.  The other adds detection of
back-to-back no-messages (where Maybe#(t_MSG) is false) and merges the
no-messages.

Two parameters control the compression:  APORT_COMPRESS_ENABLE is a static
parameter that controls whether any compression is performed.  Default is off.
APORT_COMPRESS_TIMEOUT is a dynamic parameter controlling the number of
FPGA cycles to wait for another no-message to merge after seeing a no-message.
When the default value of APORT_COMPRESS_TIMEOUT is 0 the compression mode
is statically set to Maybe#(t_MSG) compression only, with no option for
merging no-messages.